cleanup(cli): delete unused SoundLoader and duplicate FileNotFoundError
Replaced by soundpack.SoundpackResolver in production. The cli-package
FileNotFoundError shadowed soundpack.FileNotFoundError, inviting
type-assertion confusion.

Resolves review finding #7.
